redirect user to dynamic page
